#30 use custom list view to show addresses better
