+++ app/build.gradle
@@ -60,6 +60,8 @@ dependencies {
     implementation 'androidx.navigation:navigation-ui:2.3.1'
     implementation 'de.hdodenhof:circleimageview:3.1.0'
     implementation 'com.squareup.picasso:picasso:2.71828'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
 
     // Declare local unit testing dependencies
     testImplementation 'junit:junit:4.13.1'
